Move Shape Inference Pass before Constant Propagation
Before this change running "onnx-mlir --EmitONNXIR" on this pattern:

%c = "onnx.Constant"() {value = dense<"..."> : tensor<2x2xf32>} : () -> tensor<2x2xf32>
%t = "onnx.Transpose"(%32) {perm = [1, 0]} : (tensor<2x2xf32>) -> tensor<*xf32>

would result in this cast:

ShapedType resType = resOperand.getType().cast<RankedTensorType>();

in ConstPropTranspose to fail.

This is because the resOperand (the output of onnx.Transpose)
is an UnrakedTensor at this point.

This failure is fixed by moving the Shape Inference transformation to
run before Constant Propagation.