[Apollo] Disable mem-pool passes under the -npu flag.
Conditionally disable both mem-pool passes for the Apollo pipeline. Introducing the memory pool leads to having krnl.getref() statements. Such statements are later lowered in krnl -> llvm pass. At that point Memrefs are converted to a low-level representation which is incompatible with the high-level view required for turning buffers to NEPAL ArrayRefs. Disabling the passes leaves the original buffer allocations, which can subsequently be optimized and turned into NEPAL ArrayRefs in a straightforward manner.
No optimization is currently performed on the memory pools by the infra, so no performance regression is introduced.