[Apollo][BERT] Update ONNX.Unsqueeze to support bfloat16
This change updates the ONNX.Unsqueeze operator to version 13 in order to support bfloat16.

The ONNX.Unsqueeze operator also changed between versions 11 and 13 by transforming the information used to denote where to insert new dimensions from an array attribute into an argument. The majority of the changes are updating existing rewriting rules/optimizations to work with the updated operator format.

Related work items: #324