Support Expanding LayerNormalization with Dynamic Shapes
# Support Expanding LayerNormailzation with Dynamic Shapes

When importing `LayerNormalization` we need to handle the case when the input tensor has a dynamic shape. This can happen when importing from an ONNX graph (since we expand before shape inference has run) or in the case where the shapes are truly dynamic. I found this issue while trying to import BERT as an ONNX graph (previous test cases have been done using MLIR input only).

There are basically two fixes here: first, don't try and build the output type unless the input type is ranked; if the input type is fully dynamic, just use that as the output type. Second, when the type is dynamic we don't know how big to make the the `epsilon` vector, so just always create it as a single element vector and allow automatic broadcast to fix it up.