Add missing dependency of check-onnx-lit on onnx-mlir
All of the CI builds (and in our normal workflows) generally build onnx-mlir first and then proceed to build the tests, so onnx-mlir usually exists which is why this hadn't been caught before. The Windows CI build was not running the tests correctly because of python issues (see #590), so it thought they always passed marking the Windows CI build as succeeded because it is the last step to run. However, that was masking the fact that the build of onnx-mlir is failing!

This command:

```call cmake --build . --config Release --target onnx-mlir -- /m```

would run and fail. Then this command:

```call cmake --build . --config Release --target check-onnx-lit```

would run but not execute the tests and mark the build as succeeded. Since check-onnx-lit doesn't have a dependency on onnx-mlir, it does not build onnx-mlir and the earlier failure of the onnx-mlir build gets lost.

Long story short, check-onnx-lit has a dependency on onnx-mlir which this change adds. After the addition of the dependency, the Windows CI build is now correctly reporting that the Windows build of onnx-mlir in master is broken.

Signed-off-by: Stella Stamenova <stilis@microsoft.com>