Remove googletest
googletest is currently not used directly by onnx-mlir anywhere. There's a single function defined in test/unit/CMakeLists.txt that is not utilized.

Rather than adding googletest as a submodule and then adding custom defines, if any tests based on googletest are to be included in the future, they should leverage the infrastructure (add_unittest) exposed by LLVM through AddLLVM.cmake. The onnx-mlir build strictly depends on a build of LLVM, so it can be guaranteed to have access to add_unittest.

Signed-off-by: Stella Stamenova <stilis@microsoft.com>