Generate files per configuration & refer to them as such
This is a step towards fully supporting multi-config generators such as VS and Xcode. Right now, the generated files as well as path to files in the project point to a single location which often does not exist unless CMAKE_BUILD_TYPE was specified on the command line which defeats the purpose of multi-config.

This change does a few things:
1) Files that are generated by the build now reside in the config directory when using a multi-config generator. For example, instead of: src/ExternalUtil.hpp, when using VS or Xcode, we have: src/Debug/ExternalUtil.hpp, etc.
2) The targets that use the generated files now have an explicit dependency
3) The lit configuration now correctly refers to the binaries in the onnx-mlir build tree per configuration
4) The paths to onnx-mlir and the build directories are correct in the generated files
5) Executables now have the right extensions when they are referenced in tools and scripts
6) A couple of unnecessary variables are removed

Bonus: move the CTest includes where they are needed

After this change, the path to LLVM_PROJ_BIN is still not correct always since it still refers to CMAKE_BUILD_TYPE which is not always known at configuration time. This can be fixed by using find_package(LLVM) or find_package(MLIR). Using find_package is optimal, but it is a bigger change, so I am splitting it from the file generation for simplicity. The find_package(MLIR) change will be sent for review after this change (and maybe a couple more simple ones) is in.

Signed-off-by: Stella Stamenova <stilis@microsoft.com>