Add cmake tablegen functions
mlir (and llvm for that matter) use helper tablegen functions to simplify logic that is repetitive.
This is using the existing onnx_mlir_tablegen function to add a couple of helper functions similar to mlir and then use them.

As part of this change, a couple of inconsistencies in the naming of the generated targets and files were found and addressed as well.

Eventually, onnx_mlir_tablegen will replaced with mlir_tablegen (eventually = when using find_package(mlir)).

Signed-off-by: Stella Stamenova <stilis@microsoft.com>