Add the /MD flag to release builds on Windows
Without this flag, when we use ninja to build, the build will default to /MT and then we will fail to link against llvm-project. The debug build already sets /MDd

The correct thing to do, rather the most correct thing to do, would be to include HandleLLVMOptions, but we cannot do that until the circular dependencies in the project are resolved - unless we do it just in our branch and never build dynamic libraries.