[onnx] Update ops to ONNX 1.8.1
This is a temporary fix to update ONNXOps.td.inc and get check-doc passing. There is still a fundamental issue with how the build finds pybind11, that will need to be resolved in a subsequent PR.

Also un-reverts and merges !814 now that ONNX 1.8.1 opset is fully supported, required for check-doc. The primary impact of this is support for the v13 Unsqueeze op.