add_onnx_mlir_library and add_onnx_mlir_executable
Add two cmake functions to simplify the addition of libraries and executables in onnx mlir. This is based on the add* functions in llvm/mlir. The goal is to capture commonality between the setup for the various libraries/executables in onnx mlir and simplify the cmake management. There are many things that can be added to the functions but I started with a couple of simple ones that strictly improve the code without introducing too many changes:
1) The add* functions add the default include directories rather than each library having to add them. The function can be changed not to do that or to add them as private includes, but at the moment it does it for each library (since each library already did) and it adds them as public - this way any executables or tests in the project don't have to. If onnx mlir is shipped as a package in the future and the headers are installed, this can be modified to have a different set of include directories for the package interface and the change will only need to happen in the one place.
2) All libraries (unless they opt out) are added to a single global property ONNX_MLIR_LIBS. Once the libraries use target_link_libraries rather than add_dependency to manage their dependency, this global property can replace the static list OMLibs. This cannot happen yet because the libraries are added to ONNX_MLIR_LIBS in the order in which they are declared which is not the order of dependencies needed by the linker (on Linux, Windows is happy).
3) All libraries and executables are installed (unless they opt out).

Lastly, this change removes the header files from the add_library calls. They are not needed since tey are included in the cpp files and they are generally ignored by add_lirary

Signed-off-by: Stella Stamenova <stilis@microsoft.com>