Support Transpose and Reshape as High-Level Operations
# Support Transpose and Reshape as High-Level Operations

This change to onnx-mlir has onnx.reshape and onnx.transpose lowered to the `linalg.reshape` and `linalg.transpose` respectively. Previously these operations were lowered to a sequence of copy operations.

This change requires a corresponding change to the MLIR back-end to lower the high-level operations appropriately for Apollo. See !1025

## Testing
I've add a simple unit tests to verify these expansions with regard to the BERT reshapes/transposes this work is focusing on.
These changes include a small tweak to an unrelated test to catch more potential bugs in the C++ emitter.