Add Support for Reshape of Dim Size 0
# Add Support for Reshape of Dim Size 0

The BERT model's we're generating use an ONNX feature where dim of 0 means copy from input dim at that index. I originally hadn't implemented this in my Reshape code, so adding it now. There's no test for this as it needs to come form the ONNX graph (tests from graph are TBD).