Add Missing Pass
Add Missing Pass

I'm adding an affine-lowering pass to onnx-mlir. I noticed we have this pass in the maia.py script, but not in the onnx-mlir driver. I found this issue because the Transpose/Reshape code has a dependency on this pass.

## Testing
No tests yet for this as we don't have a test for the driver itself