Fix BFloat16 Constants in ONNX-MLIR
I discovered that we're not importing bfloat16 constants correctly from ONNX graph in onnx-mlir. The code was pulling the wrong 16-bits of the stored float32 value. My change fixes this bug and update or MNIST test with the corrected output.

As part of this change i'm also updating the MNIST test to use a constant that has already been rounded to be bfloat16. This ensures the model gives the same result when run in float32 mode.