HandleLLVMOptions and library dependencies
This change adds HandleLLVMOptions to the cmake file and then makes the necessary changes to support it. More specifically:
1) HandleLLVMOptions enforces that shared libs be able to link. So for each of the OM libraries, add the correct linking dependencies
2) To make 1) easier, extend the add_onnx_mlir_* functions to support dependencies, link libraries and include directories
3) Once all the OM libs link correctly, we no longer need to support a manual OMLibs list - we can rely on the global list created by add_onnx_mlir_library for linking
4) Similarly, once all the OM libs link correctly, we no longer need to support a manual list of MLIRLibs - they are all added to the appropriate OMLibs or executables instead and propagated elsewhere
5) Since onnx uses exceptions, using HandleLLVMOptions requires that we set LLVM_REQUIRES_EH to ON so that fno-exceptions is not set for onnx-mlir
6) Once all the dependencies are resolved, the custom IncGen targets for krnl and onnx are no longer needed
7) We no longer need to set /bigobj since HandleLLVMOptions does this. The same is true for /Md

A couple of caveats: there's an issue with assertions with llvm, clang and protobuf. This change sets LLVM_ENABLE_ASSERTIONS to OFF when the compiler is clang to work around it. This change also does not touch the Runtime/JNI targets. They both need some additional attention so that they can build correctly on both Windows and Linux, so they have been left alone in anticipation of getting some of those issues fixed.

Signed-off-by: Stella Stamenova <stilis@microsoft.com>