MAIA Test Framework v0.5
This is the initial check-in of the MAIA test framework i've been putting together. It still in a very early state, but its functional enough to replace the existing `ApolloSDK/MNIST` test. The framework builds on the work @<Rui Zhang> did on the `ApolloSDK/MNIST` test to generalize it.

Documentation needs to be added on how to use the framework (augmenting/replacing the existing MNIST documentation).

There are associated azure pipeline changes needed which will go in just before this PR completes. I've validated the two changes together privately.

## General Usage

The new test framework is contained in the `test/maia` directory of the onnx-mlir repo. All of the tests are built under the target `MAIA_TEST` so `ninja MAIA_TEST` builds the tests. The test are run via ctest and all of the tests are prefixed with `maia-` so the command `ctest -R maia` will run all the tests that are generated. Similarly, execution tests use the prefix `maia-execute` and the diff tests (NYI, see below) will use the prefix `maia-diff`.

## Adding Tests

All tests are dynamically generated from the script `make_tests.py`. To add a new test follow the directions in the comments of this script.

## Current Tests

There is a new test that replaces the existing MNIST test.
There are three new 'execution' tests that exercise the reshape/transpose subgraphs from the self-attention layer in BERT.
The full BERT graph tests are disabled due to compiler issues.
The generic transpose tests are disabled due to compiler isues.

## Limitations

- Model can be generated dynamically
     - static models are NYI
- Tests can be build-only, execute, diff, or a both execute and diff.
     - Diff tests are NYI
- Execute tests specify how their inputs are generated
   - static weights/bias are NYI
   - large models could cause disk-space issues; solving this is NYI
   - long running models might cause an issue; solve this is NYI
   - validating training is NYI
- Adding new tests/modifying tests
   - adding a new test or modifying how a test runs doesn't properly rerun cmake
   - please touch the CMakeLists.txt in the `maia` directory to refresh your tests
- Supports both Apollo and Athena
   - Athena support is NYI
   - I tried to isolate the Apollo specific code to make supporting Athena easier
- The Apollo SDK CMAKE functions determine input and output locations implicitly
    - This prevents me from putting the Apollo specific files into their own directory (worked around with the `Apollo.` prefix
    - Updating them to be explicit will fix this and a few other oddities in or CMAKE