Fix Race Condition in Generated Host Code
The host code were generating for our execution tests has a race condition: the `Device::WriteToHbm` ABI is non-blocking meaning that when we begin execution of a model there's no assurance that the data has finished being written to the device.

The real fix is to have the kernel wait on a notification for reach piece of data it wishes to read. This will require an ABI change to bring the semaphore from the host to the device along with the data pointer.

For now, i'm having the generated host code block on each piece of data it copies to the device before executing any code on the device. This is inefficient, but should ensure correctness.

The generated code looks like this:
```C++
  // send the data to the device
  int16_t arg0_wait = 900;
  device.WriteToHbm<npufloat_t>(arg0_address, arg0_data, Apollo::Primitives::DataType::BFloat16, -1, arg0_wait);
  int16_t arg1_wait = 901;
  device.WriteToHbm<npufloat_t>(arg1_address, arg1_data, Apollo::Primitives::DataType::BFloat16, -1, arg1_wait);
  int16_t arg2_wait = 902;
  device.WriteToHbm<npufloat_t>(arg2_address, arg2_data, Apollo::Primitives::DataType::BFloat16, -1, arg2_wait);
  int16_t arg3_wait = 903;
  device.WriteToHbm<npufloat_t>(arg3_address, arg3_data, Apollo::Primitives::DataType::BFloat16, -1, arg3_wait);
  int16_t arg4_wait = 904;
  device.WriteToHbm<npufloat_t>(arg4_address, arg4_data, Apollo::Primitives::DataType::BFloat16, -1, arg4_wait);

  // WORKAROUND: not yet passing the argument semaphores to device, so wait here to ensure copy completed
  device.ReadFromHbm(arg0_address, compute_size(arg0_dims), Apollo::Primitives::DataType::BFloat16, arg0_wait, -1);
  device.ReadFromHbm(arg1_address, compute_size(arg1_dims), Apollo::Primitives::DataType::BFloat16, arg1_wait, -1);
  device.ReadFromHbm(arg2_address, compute_size(arg2_dims), Apollo::Primitives::DataType::BFloat16, arg2_wait, -1);
  device.ReadFromHbm(arg3_address, compute_size(arg3_dims), Apollo::Primitives::DataType::BFloat16, arg3_wait, -1);
  device.ReadFromHbm(arg4_address, compute_size(arg4_dims), Apollo::Primitives::DataType::BFloat16, arg4_wait, -1);
```

## Testing
previously every few local runs of the execution tests would see one or more fail randomly with "read from uninitialized HBM".
I haven't seen any such failures since applying this fix.