Update Apollo ABI -- Input/Output Notifications
This change updates the Apollo ABI to pass notifications for each input and output tensor. Before reading from an input tensor the TCP kernel waits on its associated notification. Similarly, for output tensors, the TCP signals the associated output notification upon write completion.

The ABI change impacts each layer of the model code: the TCP kernel, the TCP interface, the DCP kernel, the DCP interface, and finally the host code. Because the host code generation, along with the execution tests, are located in this repo, while the bulk of the change are in the llvm-project repo, this will be a two part check-in (i.e. this change requires the changes to the llvm-project repo).

## ABI Details

- The host allocates a semaphore for each input and output tensor argument.
  - NEPAL doesn't support host allocation of `Notification` objects, so just use a raw integer
- The host code uses the proper APIs to associate these semaphores with the data copies on and off of the device.
- The host passes the semaphores to the DCP kernel in an array in the argument struct that sits just after the tensor arguments.
   - Input semaphores are placed first, followed immediately by output semaphores
- The DCP kernel passes the semaphores to the TCP kernel
    - The DCP kernel creates actual Notification objects from the raw semaphores
    - The DCP kernel wraps each Notification object in a singleton NotificationList
    - The TCP argument struct contains an entry for each notification with the input notifications preceding the output notifications
- The TCP kernel receives the input and out notifications as normal parameters that follow the tensor parameters
- Before reading an input tensor the TCP kernel must wait on its associated notification
- After writing to an output tensor the TCP kernel must signal its associated notification

## Implementation Details

The host code now allocates semaphore ids for each input and output tensor. These semaphores are allocated sequentially from 900. This is the recommend range for this type of semaphore. The previous workaround is removed.

I've also made two other changes
  - the host code now verifies the test is run with exactly one parameter. This fixes an AV when the test is called with no parameters.
  - i've added an explicit dependency on `mlir-opt.exe` to the cmake for the test compilation. This fixes an issue where the maia.py script isn't rerun after `mlir-opt.exe` is updated.

## Testing
MLIR tests are updated with this change
i've tested this change manually across the two repros