Implement compileModuleToSharedLibrary on Windows (#755)
* Implement compileModuleToSharedLibrary on Windows

compileModuleToSharedLibrary and friends currently do not support building on Windows - all of the default flags and properties do not work with the Windows build infrastructure. This change implements compileModuleToSharedLibrary (and any necessary functions) on Windows.

cruntime had to be updated to only depend on .c files since the .cpp file that was part of the library was not being built on Windows resulting in missing definitions when linking.

The numerical tests are also updated to work on Windows and two new build targets have been added: numerical (to build all numerical tests) and check-numerical (to run all numerical tests).

This change touches some of the JNI functions in MainUtils, but it is not sufficient to make them work on Windows. They've only been updated as a side-effect of changes to function common to both JNI and shared libs.

Signed-off-by: Stella Stamenova <stilis@microsoft.com>

* Run clang-format

Signed-off-by: Stella Stamenova <stilis@microsoft.com>

* Add description for add_numerical_test

Signed-off-by: Stella Stamenova <stilis@microsoft.com>

Co-authored-by: Kevin O'Brien <caomhin@us.ibm.com>
Co-authored-by: gongsu832 <gong_su@hotmail.com>