Use the LLVM_HOST_TRIPLE as the default for MainUtils if the triple exists
This allows every executable that links against MainUtils (such as onnx-mlir, tests, etc.) to default to the host triple. That way if the build of llvm used a different default triple (that is often not supported by onnx mlir), it would not be the default for onnx mlir and the tests (allowing the tests to pass on the platform where onnx mlir was built). If -mtriple is specified, it still takes precedence to the default.

Signed-off-by: Stella Stamenova <stilis@microsoft.com>