Move SimplifyKrnl.cpp to src/Dialect/Krnl/ to avoid cyclic dependency.
After a patch D105903 was merged in MLIR, got link-time failure in shared
library build. This is because the patch forces the compiler to emit a
strong definition for a class. To avoid this, OMKrnlOps need to be added
in LINK_LIBS of OMSimplifyKrnl. However, this introduces cyclic
dependency(OMSimplifyKrnl -> OMKrnlOps -> OMSimplifyKrnl). So, this
commit moves SimplifyKrnl.cpp in OMKrnlOps to avoid cyclic dependency.

Reference: https://reviews.llvm.org/D105903

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>