Use the LLVM_HOST_TRIPLE as the default for MainUtils if the triple exists (#793)
* Use the LLVM_HOST_TRIPLE as the default for MainUtils if the triple exists

This allows every executable that links against MainUtils (such as onnx-mlir, tests, etc.) to default to the host triple. That way if the build of llvm used a different default triple (that is often not supported by onnx mlir), it would not be the default for onnx mlir and the tests (allowing the tests to pass on the platform where onnx mlir was built). If -mtriple is specified, it still takes precedence to the default.

Signed-off-by: Stella Stamenova <stilis@microsoft.com>

* Make sure that get_host_triple can find config.guess on platforms that use it

Signed-off-by: Stella Stamenova <stilis@microsoft.com>

* PR feedback

Signed-off-by: Stella Stamenova <stilis@microsoft.com>

* PR feedback:
Add ONNX_MLIR_DEFAULT_TRIPLE which is a cmake property that can be set at compilation time to choose the default triple.
Rename kHostTriple to kDefaultTriple since that better represents the intention.
Add information to the readme about the options for setting the triple.
Remove mtriple and mcpu from test.py since they both default to the onnx-mlir defaults and the tests should pass with those settings.

Signed-off-by: Stella Stamenova <stilis@microsoft.com>

* Revert changes to test.py

Signed-off-by: Stella Stamenova <stilis@microsoft.com>

* The changes required to use the default llvm host triple in all cases were committed to llvm-project a while back, so after the llvm merge we can rely on them

Signed-off-by: Stella Stamenova <stilis@microsoft.com>

Co-authored-by: Kevin O'Brien <caomhin@us.ibm.com>
Co-authored-by: Alexandre Eichenberger <alexe@us.ibm.com>