We currently run a few test sets, e.g., check-onnx-backend, check-onnx-backend-dynamic, etc., in serial. While we do parallelize within each test set, we don't get much benefit from it. Because within each test set, most of the time was spent on one or two long running tests. While waiting for these long running tests, we could have started the other test sets.
This patch parallelizes across tests sets by starting them
all at once so the long running tests from different test
sets can run in parallel.

CompilerLib tests are disabled because a bug has been uncovered
that it was actually still using onnx-mlir instead of CompilerLibTest
to compile the models. Fix the bug to use CompilerLibTest and
it fails to compile the models.

Use CTEST_PARALLEL_LEVEL for numerical tests to be compatible
with Windows CI

Signed-off-by: Gong Su <gong_su@hotmail.com>