Force check-doc to use Python3_EXECUTABLE (#1193)
In our team's subtree-merge of onnx-mlir, we configure a venv to give us precise control over the python configuration without messing with the system/user python install. The check-doc target fails in this mode, because it defaults to the python in the path. There are a couple of ways to resolve this: one would be to run the tests with 'cmake -E env' and inject the python dir at the head of the PATH. This has insidious potential side effects, as the test will always run with the value of PATH as it was at configure time. Subsequent changes to the PATH in a shell window will have no effect without rebuilding the cmake cache, and then those changes will persist until the next cache rebuild.

The other approach, which is what I've implemented here, is to render the doc-check file as cmake templates via configure_file. While slightly more complex from an implementation point of view, it is more robust and the results more predictable.

Signed-off-by: Henry Morgan <hmorgan@xbox.com>

Co-authored-by: Henry Morgan <hmorgan@xbox.com>
Co-authored-by: Stella Stamenova <stilis@microsoft.com>
Co-authored-by: gongsu832 <gong_su@hotmail.com>
Signed-off-by: Ethan Wang <ywan2928@uwo.ca>