Add the llvm source version to VCSVersion.inc even if we don't have access to the llvm source (#1248)
* Add the llvm source version to VCSVersion.inc even if we don't have access to the llvm source

Right now the version is only appended when the source is explicitly available. This is not always the case since onnx-mlir can be built against a build directory as well as against an install directory of llvm. This change tries to check for the llvm source version and if it not found, includes the VCSRevision.h header generated by llvm instead.

Signed-off-by: Stella Stamenova <stilis@microsoft.com>