Fix output path verification for onnx-mlir and update the lit tests to not pollute the source tree (#1329)
Some lit tests leave artifacts in the source tree. Update the tests, so that all artifacts are placed in the binary tree.

This revealed a bug in the output path verification (checking whether the specified output path is valid), so the verification is updated as well. It now works correctly regardless of the type of slashes used as well as regardless of the depth of the paths.

Signed-off-by: Stella Stamenova <stilis@microsoft.com>