Derive a new max trip count for ONNXLoopOp when the upperbound, lowerbound and step are not constants (#1462)
Before this commit, we have a rewriting rule to derive a new max trip count for ONNXLoopOp when the upperbound, lowerbound and step are constants. In such case, the new max trip count is computed during the compile time.

This commit enhances the rule to support the case wherein the upperbound, lowerbound and step are not constants. In this case, we emit code to compute a derived trip count, and the new max trip count will be min(old_max_trip_count, derived_max_trip_count).

Signed-off-by: Tung D. Le <tung@jp.ibm.com>