Optimized symbol table lookup API
Instead of first calling "contains" and only if true call "get",
now we just call "get" which returns a pointer and signals that
the key is not found by returning nullptr. This halves the number
of lookups in the underlying nameToObject map.

Also changed the way EmitInitializerForInputTensor is called
which avoids a copy of the initializer tensor.

Signed-off-by: Soren Lassen <sorenlassen@gmail.com>