build: decouple CompilerOptions from ONNXToKrnl pass (#1671)
Prior to this patch, any consumer of the ONNXToKrnl pass (transitively)
depended on the CompilerOptions library.  This dependency caused
conflicts in identically-named command line options between the
CompilerOptions library and the consumers of the ONNXToKrnl pass.

This patch introduces a local option to the ONNXToKrnl pass so that its
dependency on CompilerOptions is removed, thus eliminating the symbol
conflicts.  When its time to pass a command line option to enable
parallelization in the ONNXTOKrnl pass, the call to
`createLowerToKrnlPass()` in CompilerPasses.cpp, which correctly depends
on the CompilerOptions library, will be able to pass the command line
value to the ONNXToKrnl pass.

Signed-off-by: Ashay Rane <ashay@users.noreply.github.com>

Signed-off-by: Ashay Rane <ashay@users.noreply.github.com>
Co-authored-by: Tung D. Le <tung@jp.ibm.com>