avoid re-definition of macro (#1672)
This patch adds a guard statement to only define the
`_SILENCE_ALL_CXX17_DEPRECATION_WARNINGS` macro if it has not previously
been defined.  Similarly, it checks to make sure that the macro has been
defined before un-defining it, but only if it was defined in this file.

This patch is necessary if there is a dependency on ORT or another third
party library that sets the macro.  If there is no such dependency, then
the macro is correctly set and unset in the root CMakeLists.txt, making
this patch a NOP.

Signed-off-by: Ashay Rane <ashay@users.noreply.github.com>