Hybrid ONNX-to-ONNX transformation pass
Includes a new pattern-based shope inference implementation.

So far only combines shape inference and canonicalization but
the plan is to include constant propagation and decomposition
so that a single pass with all the patterns can cascade
shape inference and all the transforms in a single pass.

I encountered a model with a long dependency chain between
shape inference, canonicalization, and constant propagation
which requires an incredible number of repetitive passes
with the current pass infrastructure.

Signed-off-by: Soren Lassen <sorenlassen@gmail.com>