Fix RuntimeAPIRegistry singleton class memory leak.
The singleton class pointer was set to null between calls to compileModule(),
which leads to memory leak. The idea of a singleton is that it should last for
the lifetime of the process. However, the fact that it cannot be shared across
compileModule() suggests that it's not the right construct to use. In addition,
the singleton calss is not thread safe without proper serialization.

This patch changes the singleton to a normal class. It's verified to fix the
memory leak reported by valgrind and also makes the code thread safe.

Signed-off-by: Gong Su <gong_su@hotmail.com>