Support faster sort algorithm in onnx.NonMaxSuppression and onnx.TopK (#1869)
This PR introduces a faster sort algorithm (quick sort and comb sort) on CPU in the emitArgSort function.
The quick sort algorithm is used for onnx.NonMaxSupression and onnx.TopK operations.

[Conditions to use faster sort algorithm]
This PR uses faster sort algorithms if the following two conditions are met. The original bubble sort code is used in other cases.

    the input rank should be less or equal to 6
    the axis to be sorted should be the last rank. Otherwise the original bubble sort code is used.
    I consider that the conditions are met in all onnx.NonMaxSupression cases and the most practical cases of onnx.TopK.

[How to guarantee stable sort]
This function is expected to provide "stable" sort (e.g. "test_nonmaxsuppression_identical_boxes_cpu" test in backend test), so that it should preserve the input data order, if the values are the same, although most sort algorithms (e.g. quick sort) are not stable.
To guarantee the original order among the same values, we use special compare functions to compare two values at first, then compare the input orders if the values are the same. Luckily the target list is managed by index and data arrays, and the index array is initialized by the input order (1, 2, 3, ...), so that we can compare the input orders as well as their values.

Signed-off-by: Yasushi Negishi negishi@jp.ibm.com