onnx-mlir: [nfc] create target machine using triple, cpu, and arch
In preparation for a subsequent patch, which will drop the `--mtriple`,
`--march`, and `--mcpu` flags in favor of library invocations of the
LLVM optimizer and assembler, this patch sets up the LLVM target machine
using the provided values of the triple, CPU, architecture.  Setting up
the LLVM target machine this way avoids the need to update the triple,
CPU, and architecture later.

Signed-off-by: Ashay Rane <ashay@users.noreply.github.com>