onnx-mlir: replace binary invocations of opt and llc with library calls
Prior to this patch, onnx-mlir and onnx-mlir-opt invoked the opt and llc
binaries, which required these binaries and all LLVM libraries to be
copied to the same directory where onnx-mlir artifacts are installed.
This also created problems in shared library builds, since copying the
binaries and libraries does not update the RPATH embedded in them.

To resolve these problems, this patch replaces invocations of the opt
and llc binaries in CompilerUtils.cpp with calls to LLVM APIs for
running the optimizer and the assembler.  As a minor ancillary benefit,
this change allows us to skip writing the unoptimized bitcode file to
disk, since the same `llvm::Module` object can be carried over to the
optimizer code as well as the assembler code.

A downside of this change is that we can no longer pass arbitrary
options to opt and llc, like we previously could using the `Xopt`,
`Xllc`, and `mllvm` flags to onnx-mlir and onnx-mlir-opt.

Signed-off-by: Ashay Rane <ashay@users.noreply.github.com>