Add a requirements.txt file for PyPi requirements in onnx-mlir (#2010)
This adds a `requirements.txt` file at the root of the repository with the current PyPi package requirements and pinned versions for them. This is the equivalent of the mlir change that pinned their package requirements (https://reviews.llvm.org/D142563).

This change is pinning the requirements to a specific version (or a range) depending on the requirement. A couple of considerations:

* numpy 1.24 deprecates np.object, np.bool, np.float, np.complex, np.str, and np.int which are used heavily in onnx-mlir
* not all versions of each package are available on every platform - to the best of my knowledge, these ranges should work on Ubuntu, CentOS, macOS and Windows

Adding a minimum and maximum version, or pinning to a specific versions where possible, helps with two major goals - security and maintainability. It gives us an opportunity to make sure that the packages being used are not part of a security attack as well as guaranteeing that they support the features that onnx-mlir depends on (see note about numpy deprecation).

This also adds the `requirements.txt` file to the setup instructions and to the pipelines on Windows and macOS. I've not added it to the docker images but in theory it may be possible to use it there as well.

Signed-off-by: Stella Stamenova <stilis@microsoft.com>