Make the doc example test work with Ninja on Ubuntu (#2023)
This should, in theory, work on all platforms now.

Before this change, the test worked correctly with Makefiles on any platform because Makefiles always use the full path to the library. It also worked correctly on Windows regardless of generator because by default an executable will look for any dlls in the same directory first. However, it did not work correctly with ninja on non-Windows because ninja was converting the full path to a path relative to the root of the binary tree for the build step and then failing at execution. More specifically, the library would be linked as docs/doc_example/add.so, then at execution time of the test, the test would run from docs/doc_example, so it would look for the library in docs/doc_example/docs/doc_example. The behavior could not be altered by setting rpath or by explicitly passing the full path (without using the imported library).

To make the test work everywhere, I've changed the output library to be add on Windows and libadd on non-Windows. This allows us to set the IMPORTED_NO_SONAME target property which will then link the library via -ladd instead of by relative path.

Signed-off-by: Stella Stamenova <stilis@microsoft.com>