mhlo: migrate from MLIR-HLO to StableHlo
This patch replaces the ONNX-to-MHLO conversion pipeline with the
ONNX-to-StableHlo conversion pipeline.  Aside from the relatively few
changes to the C++ source files, most of the changes in this patch
update the lit tests, all of which were updated using the
utils/mlir2FileCheck.py script.

This patch does not rename the ONNX_MLIR_ENABLE_MHLO CMake flag nor does
it make the repository point to the StableHlo submodule.  I plan to make
these changes in a separate patch, since build-related changes are
likely to take a life of their own.

During the migration, I noticed a couple of bugs in the StableHlo
generation/lowering, specifically, in the generation/lowering of MatMul
and Convolution StableHlo operations.  I have disabled these tests for
now and if/when this patch is accepted, I'll create a GitHub issue so
that we can track the fixes.

Signed-off-by: Ashay Rane <ashay@users.noreply.github.com>