mhlo: migrate from MLIR-HLO to StableHlo
This patch replaces the ONNX-to-MHLO conversion pipeline with the
ONNX-to-StableHlo conversion pipeline.  Aside from the relatively few
changes to the C++ source files, most of the changes in this patch
update the lit tests, all of which were updated using the
utils/mlir2FileCheck.py script.

This patch also makes this repository point to the StableHLO submodule
(and drops the MLIR-HLO submodule) since our MHLO commit was old and
contained build bugs that were fixed upstream in the StableHLO
repository.

During the migration, I noticed that MatMul operations with dynamic
dimensions are not lowered correctly.  Specifically, the lowering for
matmul operations includes StableHLO's BroadcastInDimOp, which requires
that its dimensions be static.  Fixing this requires a slightly more
involved fix, but I will file an issue for this and fixing it in a
separate PR.  For now, I have inserted a guard condition that skips the
lowering for matmul with dynamic dimensions, so that the pass doesn't
crash.

Signed-off-by: Ashay Rane <ashay@users.noreply.github.com>