feat(ONNXOps): add manually bfloat16 support
for certain operations while we wait for upstream support. The current
list of operations are: Conv, MaxPool, AveragePool and GlobalAveragePool

These manually changes should be removed once the upstream definition
has been approved and updated. This is only a temporary solution for now
to get a bfloat16 flow to run with CNNs.