[Cherry-pick] Fixing the location of DimAnalysis in onnx-to-zhigh pass and some rules in zhigh-to-onnx pass
The onnx-to-zhigh pass has two phases: 1) converting multiple onnx ops into a single zhigh op, and 2) converting a single onnx op to a single zhigh op, where the second phase uses DimAnalysis (Patterns in the 1st phase at this moment does not use DimAnalysis)

The problem is DimAnalysis is currently called before the 1st phase, which is not good because the 1st  phase may change the IR so the information from DimAnalysis is obsoleted to the 2nd phase. Correct position for DimAnalysis would be just before the 2nd phase.

Other than that, this PR changes slightly the rules in zhigh-to-onnx pass so that for binary ops, only one input (instead of two) that is from stick would be enough to trigger the rule to convert a zhigh op back to an onnx op.

Resolves #2789

---------

Signed-off-by: Tung D. Le <tung@jp.ibm.com>
(cherry picked from commit 80a63f2c1f497f50c1f3ebfb3372dfc3ef50ff2d)
Signed-off-by: Charles Volzka <cjvolzka@us.ibm.com>